requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo
pymongo[srv]
scikit-learn
matplotlib


##-e .
